gba: rename dkongcrest to dkongc
This matches the ROM set name with the vanilla game, and since the
save RAM data is compatible (in fact, you can play input recordings on
both without desyncs), it's kind of nice to be able to keep them
loading/saving the same data.

This idea might not apply for future enhancement hacks.  It'll both
depend on whether save data is compatible and how many enhancement
hacks are added.